chore: refactor routes to clean up responsibilities
